Add support for media types
For some endpoints GitHub lets you request different response format
using the `Accept` header [[1]]. However, by default if the response is
not in JSON format `GhApi.__call__` will raise an error.

This commit makes it possible by adding a bit of logic to look at the
`Accept` header in the request and tell `fastcore.core.urlsend` not to
return JSON if it doesn't look like the user is requesting a JSON media
type.

[1]: https://docs.github.com/en/rest/overview/media-types